Move matcher to unmock-core.
