+++ packages/unmock-core/package.json
@@ -13,7 +13,9 @@
     "url": "https://github.com/unmock/unmock-js/issues"
   },
   "dependencies": {
+    "@types/debug": "^4.1.4",
     "axios": "^0.18.0",
+    "debug": "^4.1.1",
     "querystring": "^0.2.0",
     "unmock-hash": "file:../unmock-hash"
   }
