--- packages/unmock-matcher/README.md
@@ -1,3 +1,0 @@
-# unmock-matcher
-
-Implementation of the unmock matcher.
