--- packages/unmock-matcher/package.json
@@ -1,20 +1,0 @@
-{
-  "version": "0.0.55",
-  "name": "unmock-matcher",
-  "displayName": "unmock-matcher",
-  "main": "dist/index.js",
-  "license": "MIT",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/unmock/unmock-js.git"
-  },
-  "homepage": "https://github.com/unmock/unmock-js",
-  "bugs": {
-    "url": "https://github.com/unmock/unmock-js/issues"
-  },
-  "dependencies": {
-    "@types/debug": "^4.1.4",
-    "debug": "^4.1.1",
-    "unmock-core": "file:../unmock-core"
-  }
-}
