--- packages/unmock-matcher/src/__tests__/tsconfig.json
@@ -1,7 +1,0 @@
-{
-  "extends": "../../../../tsconfig.test.base",
-  "include": ["**/*"],
-  "references": [
-    { "path": "../../" }
-  ]
-}
