--- packages/unmock-matcher/src/index.ts
@@ -1,1 +1,0 @@
-export * from "./matcher";
