+++ tsconfig.json
@@ -6,6 +6,6 @@
   "include": [],
   "references": [
     { "path": "./tsconfig.build.json" },
-    { "path": "./tsconfig.test.json" },
+    { "path": "./tsconfig.test.json" }
   ]
 }
